requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.4
billiard==4.2.1
bs4==0.0.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
kombu==5.5.3
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
timedelta==2020.12.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0